Added spikes and breakable bridges to level1
