+++ Assets/Prefabs/Interactables/BreakableBridgeTile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8eb119c0a4bb19949a3e2d406b6641cb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
