+++ Assets/Scriptable Object Instances/BreakableBridgeTile.asset
@@ -1,0 +1,39 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 690ace13da474fe2a899b9209cbb941a, type: 3}
+  m_Name: BreakableBridgeTile
+  m_EditorClassIdentifier: 
+  m_Sprite: {fileID: 21300006, guid: 84632f772d45a04488e6c5318eb12242, type: 3}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
+  m_Transform:
+    e00: 1
+    e01: 0
+    e02: 0
+    e03: 0
+    e10: 0
+    e11: 1
+    e12: 0
+    e13: 0
+    e20: 0
+    e21: 0
+    e22: 1
+    e23: 0
+    e30: 0
+    e31: 0
+    e32: 0
+    e33: 1
+  m_InstancedGameObject: {fileID: 0}
+  m_Flags: 1
+  m_ColliderType: 1
+  TileSprite: {fileID: 21300006, guid: 84632f772d45a04488e6c5318eb12242, type: 3}
+  TileAssociatedPrefab: {fileID: 1044135072517746560, guid: 8eb119c0a4bb19949a3e2d406b6641cb,
+    type: 3}
