+++ Assets/Scriptable Object Instances/BreakableBridgeTile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 728f4115b214c6443ab99bd92fd4b1b3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
