+++ Assets/Scriptable Object Instances/IronSpike.asset
@@ -9,7 +9,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 79421526d9f81174c8cb78e05d00c85e, type: 3}
+  m_Script: {fileID: 11500000, guid: 690ace13da474fe2a899b9209cbb941a, type: 3}
   m_Name: IronSpike
   m_EditorClassIdentifier: 
   m_Sprite: {fileID: -1929834416523346238, guid: e1a2afb1f8f971842a05df992837888d,
@@ -39,7 +39,3 @@ MonoBehaviour:
     type: 3}
   TileAssociatedPrefab: {fileID: 1185636220084658051, guid: 437a317da3deaee4b9729cec546c0a58,
     type: 3}
-  PrefabLocalZOffset: 0
-  UseAbsoluteZOffset: 0
-  PrefabAbsoluteZOffset: 0
-  TileIsEditorObjectOnly: 0
