+++ Assets/Scripts/Interactables/BreakableBridgeTileBehaviour.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: ec3d819572134bec8a698905ef4fcff2
+timeCreated: 1606093917
