+++ Assets/Scripts/Interactables/ScriptableTiles/ScriptableTile.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 690ace13da474fe2a899b9209cbb941a
+timeCreated: 1606093166
