--- Assets/Scripts/Interactables/ScriptableTiles/SpikeTile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 79421526d9f81174c8cb78e05d00c85e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
