+++ Assets/Scripts/WeaponsSystem/MonoBehaviours/ThrowableBehaviourScript.cs
@@ -87,9 +87,9 @@ namespace MonoBehaviours.WeaponsSystem
                     var playerController = hit.GetComponent<Controller>();
                     playerController.LoseHitPoints(_projectileData.ProjectileBaseDamage);
                 }
-                else if (hit.CompareTag("ExplosiveBarrel"))
+                else if (hit.GetComponent<IBreakable>() != null)
                 {
-                    GetComponent<IBreakable>().Break();
+                    hit.GetComponent<IBreakable>().Break();
                 }
                 // todo: interacts with breakable things
             }
