+++ Assets/Tilemaps/Level.prefab
@@ -49,7 +49,16 @@ Tilemap:
       m_TileMatrixIndex: 0
       m_TileColorIndex: 0
       m_TileObjectToInstantiateIndex: 0
-      m_AllTileFlags: 4
+      m_AllTileFlags: 1073741828
+  - first: {x: -5, y: -3, z: 0}
+    second:
+      serializedVersion: 2
+      m_TileIndex: 31
+      m_TileSpriteIndex: 2
+      m_TileMatrixIndex: 0
+      m_TileColorIndex: 0
+      m_TileObjectToInstantiateIndex: 1
+      m_AllTileFlags: 1073741828
   - first: {x: -7, y: -1, z: 0}
     second:
       serializedVersion: 2
@@ -384,13 +393,15 @@ Tilemap:
     m_Data: {fileID: 11400000, guid: cc9b7fd15fa00014d8b9cf2232d6b1b7, type: 2}
   - m_RefCount: 1
     m_Data: {fileID: 11400000, guid: 35ff0df43c45ec541bc1edc096044122, type: 2}
+  - m_RefCount: 1
+    m_Data: {fileID: 11400000, guid: 728f4115b214c6443ab99bd92fd4b1b3, type: 2}
   m_TileSpriteArray:
   - m_RefCount: 1
     m_Data: {fileID: -1929834416523346238, guid: e1a2afb1f8f971842a05df992837888d,
       type: 3}
   - m_RefCount: 1
     m_Data: {fileID: 21300006, guid: 9dc18d9dbefddc44ebab3f88ec5da09c, type: 3}
-  - m_RefCount: 1
+  - m_RefCount: 2
     m_Data: {fileID: 21300006, guid: 84632f772d45a04488e6c5318eb12242, type: 3}
   - m_RefCount: 1
     m_Data: {fileID: 21300006, guid: b00b9ff8e2a7f494c9679b0b59c599bb, type: 3}
@@ -449,7 +460,7 @@ Tilemap:
   - m_RefCount: 1
     m_Data: {fileID: 21300000, guid: 78ba6269d74827a418526a8079d870d6, type: 3}
   m_TileMatrixArray:
-  - m_RefCount: 31
+  - m_RefCount: 32
     m_Data:
       e00: 1
       e01: 0
@@ -468,11 +479,16 @@ Tilemap:
       e32: 0
       e33: 1
   m_TileColorArray:
-  - m_RefCount: 31
+  - m_RefCount: 32
     m_Data: {r: 1, g: 1, b: 1, a: 1}
+  - m_RefCount: 0
+    m_Data: {r: -1.841733e+37, g: -1.841733e+37, b: -1.841733e+37, a: -1.841733e+37}
   m_TileObjectToInstantiateArray:
   - m_RefCount: 1
     m_Data: {fileID: 1185636220084658051, guid: 437a317da3deaee4b9729cec546c0a58,
+      type: 3}
+  - m_RefCount: 1
+    m_Data: {fileID: 1044135072517746560, guid: 8eb119c0a4bb19949a3e2d406b6641cb,
       type: 3}
   m_AnimationFrameRate: 1
   m_Color: {r: 1, g: 1, b: 1, a: 1}
@@ -584,7 +600,7 @@ Grid:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 7070841107302164381}
   m_Enabled: 1
-  m_CellSize: {x: 1, y: 1, z: 0}
+  m_CellSize: {x: 0.93, y: 0.99, z: 0}
   m_CellGap: {x: 0, y: 0, z: 0}
   m_CellLayout: 0
   m_CellSwizzle: 0
